Move timezone utils to dedicated file
