+++ apps/website/src/components/show-and-tell/GiveAnHourProgress.tsx
@@ -4,7 +4,7 @@ import { useMemo } from "react";
 import type { PartialDateString } from "@alveusgg/data/build/types";
 
 import { classes } from "@/utils/classes";
-import { DATETIME_ALVEUS_ZONE } from "@/utils/datetime";
+import { DATETIME_ALVEUS_ZONE } from "@/utils/timezone";
 import { trpc } from "@/utils/trpc";
 
 import useLocaleString from "@/hooks/locale";
